Collapse metrics module namespace.
