+++ examples/web_server_with_debug_port.rs
@@ -5,9 +5,7 @@ extern crate histogram;
 
 use iron::prelude::*;
 use iron::status;
-use metrics::metrics::counter::*;
-use metrics::metrics::gauge::*;
-use metrics::metrics::meter::*;
+use metrics::metrics::{Counter, Gauge, Meter, StdCounter, StdGauge, StdMeter};
 use metrics::registry::StdRegistry;
 use metrics::reporter::PrometheusReporter;
 use std::sync::Arc;
